# swagger.yaml
openapi: 3.0.0
info:
  title: Notes, Products, and Orders API
  version: 1.0.0
  description: API for user notes, secure product catalog (menu), and order processing (stored as JSON in notes).

servers:
  - url: http://localhost:9001/api
    description: Development Server Base URL

# =======================================================
# PATHS (Endpoints)
# =======================================================
paths:
  # --- 1. AUTH ROUTES ---
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user account.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "201":
          description: User created successfully.
        "400":
          description: Validation error or username already exists.

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Log in and receive a JWT token.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Successful login, returns JWT in the response body or cookies.
        "401":
          description: Invalid credentials.

  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get the current user's info (ID, username, role).
      description: Returns the user object populated by the JWT token. Used for checking roles.
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Successfully retrieved user info.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
                  role:
                    type: string
                    example: admin
        "401":
          description: Unauthorized.

  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Log out (typically clears JWT cookie/storage).
      operationId: logoutUser
      responses:
        "200":
          description: Logout successful.

  /auth/delete-account:
    delete:
      tags:
        - Authentication
      summary: Delete the user's own account.
      operationId: deleteSelf
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Account deleted successfully.
        "401":
          description: Unauthorized.

  # --- 2. PUBLIC PRODUCT ROUTES (MENU) ---
  /products:
    get:
      tags:
        - Products (Public Menu)
      summary: Get the list of all ACTIVE products (The Restaurant Menu).
      operationId: listPublicProducts
      responses:
        "200":
          description: A list of active products.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product" # Based on product_service.js getActiveProducts

  /products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: The ID of the product.
    get:
      tags:
        - Products (Public Menu)
      summary: Get a single ACTIVE product by ID.
      description: Returns product details ONLY if the product is marked as active (is_active = 1).
      operationId: getSingleActiveProduct
      responses:
        "200":
          description: Single active product object.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found or is inactive.

  # --- 3. CUSTOMER ORDER ROUTES ---
  /orders:
    post:
      tags:
        - Orders (Transaction)
      summary: Create a new order (stored as JSON in a Note).
      description: Requires a valid JWT. The server calculates the total price based on the current product catalog.
      operationId: createOrder
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput" # From order_schema.js
      responses:
        "201":
          description: Order successfully created and saved as a note record.
        "400":
          description: Validation failed (e.g., quantity < 1 or invalid product ID).
        "401":
          description: Unauthorized.

    get:
      tags:
        - Orders (Transaction)
      summary: Get a list of the current user's orders (filtered from notes).
      description: Fetches records from the 'notes' table that contain a valid order JSON structure.
      operationId: listUserOrders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of structured order objects.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
        "401":
          description: Unauthorized.

  # --- 4. CUSTOMER NOTES ROUTES ---
  /notes:
    post:
      tags:
        - Notes (Customer)
      summary: Create a new general text note.
      operationId: createNote
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput" # From note_schema.js
      responses:
        "201":
          description: Note created successfully.
    get:
      tags:
        - Notes (Customer)
      summary: Get list of all user's notes and orders.
      operationId: getCustomerNotes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of notes.

  /notes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: The ID of the note or order.
    get:
      tags:
        - Notes (Customer)
      summary: Get a single note or order by ID.
      operationId: getSingleNote
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Single note/order object.
        "404":
          description: Note not found or not owned by the user.
    put:
      tags:
        - Notes (Customer)
      summary: Update an existing note (not recommended for Order notes).
      operationId: updateNote
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput" # From note_schema.js
      responses:
        "200":
          description: Note updated successfully.
    delete:
      tags:
        - Notes (Customer)
      summary: Delete a note or order.
      operationId: deleteNote
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Note deleted successfully.

  # --- 5. ADMIN PRODUCT CRUD ROUTES ---
  /admin/products:
    get:
      tags:
        - Admin (Products)
      summary: Get ALL products (active/inactive) for the admin panel.
      operationId: listAllProductsAdmin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: A list of all products (requires admin role).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product" # Based on product_service.js getAllProductsAdmin
    post:
      tags:
        - Admin (Products)
      summary: Create a new product.
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product successfully created.

  /admin/products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: The ID of the product.

    get:
      tags:
        - Admin (Products)
      summary: Get a single product (active or inactive) for editing.
      operationId: getSingleProductAdmin
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Full product details (requires admin role).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized.
        "403":
          description: Forbidden (Requires Admin role).
        "404":
          description: Product not found.

    put:
      tags:
        - Admin (Products)
      summary: Update product details (full or partial update).
      operationId: updateProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Product successfully updated.
    delete:
      tags:
        - Admin (Products)
      summary: Delete a product.
      operationId: deleteProduct
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Product successfully deleted.

  # --- 6. ADMIN USER/NOTE ROUTES ---
  /admin/users:
    get:
      tags:
        - Admin (Users/Notes)
      summary: Get list of all users.
      operationId: listAllUsers
      security:
        - bearerAuth: []
      parameters: # Based on user_service.js getAllUsers
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of all users (requires admin role).

  /admin/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    delete:
      tags:
        - Admin (Users/Notes)
      summary: Delete a specific user.
      operationId: deleteSpecificUser
      security:
        - bearerAuth: []
      responses:
        "204":
          description: User deleted successfully.

  /admin/notes:
    get:
      tags:
        - Admin (Users/Notes)
      summary: Get list of all notes (and orders) from all users.
      operationId: listAllNotesAdmin
      security:
        - bearerAuth: []
      parameters: # Based on note_service.js adminGetAllNotes
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of all notes/orders in the system (requires admin role).

  /admin/notes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    delete:
      tags:
        - Admin (Users/Notes)
      summary: Delete a specific note/order by ID.
      operationId: deleteSpecificNoteAdmin
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Note/order deleted successfully.

  /admin/orders/{id}/status:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: The ID of the order (which is the note ID).
    put:
      tags:
        - Admin (Orders)
      summary: Update the status of an existing order.
      description: Allows an admin to change the order status (e.g., to READY or DELIVERED) by modifying the JSON content within the note record.
      operationId: updateOrderStatusAdmin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdateInput"
      responses:
        "200":
          description: Order status updated successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  newStatus:
                    type: string
                    enum: [PENDING, IN_PROGRESS, READY, DELIVERED, CANCELLED]
        "400":
          description: Validation error (invalid status) or the record is not a valid order.
        "401":
          description: Unauthorized.
        "403":
          description: Forbidden (Requires Admin role).
        "404":
          description: Order not found.

# =======================================================
# COMPONENTS (Schemas and Security)
# =======================================================
components:
  schemas:
    # 1. AUTH SCHEMAS (from auth_schema.js)
    AuthInput:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
          example: testuser
        password:
          type: string
          minLength: 6
          maxLength: 20
          example: password123
      required:
        - username
        - password

    # 2. NOTE SCHEMAS (from note_schema.js)
    NoteInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
          example: My first note
        content:
          type: string
          maxLength: 5000
          example: This is a general note.
      required:
        - title

    # 3. PRODUCT SCHEMAS (from product_schema.js)
    ProductInput:
      type: object
      properties:
        name:
          type: string
          minLength: 3
          example: Spicy Pepperoni Pizza
        description:
          type: string
        price:
          type: number
          format: float
          minimum: 0.01
          example: 15.99
        category:
          type: string
          example: Pizzas
        tags:
          type: string
          example: spicy, popular
        extra_info:
          type: string
          example: Chef's special
        image_url:
          type: string
          example: pizza-pep.jpg
        is_active:
          type: boolean
          example: true

    Product:
      allOf:
        - $ref: "#/components/schemas/ProductInput"
        - type: object
          properties:
            id:
              type: integer
              example: 101

    # 4. ORDER SCHEMAS (from order_schema.js & product_schema.js)
    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
          minimum: 1
          example: 101
        quantity:
          type: integer
          minimum: 1
          example: 2
      required:
        - productId
        - quantity
    OrderInput:
      type: array
      description: Array of items to be ordered.
      items:
        $ref: "#/components/schemas/OrderItemInput"
      minItems: 1

    OrderResponseItem:
      type: object
      description: Detailed item entry in the saved order.
      properties:
        productId:
          type: integer
          example: 101
        name:
          type: string
          example: Spicy Pepperoni Pizza
        price:
          type: number
          format: float
          example: 15.99
        quantity:
          type: integer
          example: 2
        total:
          type: number
          format: float
          example: 31.98
      required:
        - productId
        - name
        - price
        - quantity
        - total

    OrderResponse:
      type: object
      description: A full order record fetched from the notes content.
      properties:
        orderId:
          type: integer
          example: 50
          description: The ID of the note record that stores this order.
        title:
          type: string
          example: Order placed on 10/25/2024
        user_id:
          type: integer
          example: 2
        order_date:
          type: string
          format: date-time
        status:
          type: string
          example: PENDING
        total_amount:
          type: number
          format: float
          example: 45.97
        order_items:
          type: array
          items:
            $ref: "#/components/schemas/OrderResponseItem"
      required:
        - orderId
        - user_id
        - total_amount
        - order_items

    OrderStatusUpdateInput:
      type: object
      description: Input required to update the status of an order.
      properties:
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, READY, DELIVERED, CANCELLED]
          example: DELIVERED
      required:
        - status

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT Bearer token obtained from /auth/login.